+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [UNRELEASED]
 ### Changed
 - `CDSInterval` object now has methods to access the number of codons and codon locations in both chunk-relative and chromosome coordinates. Chromosome accessors will always return the full original CDS.
+- If duplicate sequence identifiers are found when parsing GenBank/FASTA files, an exception is raised.
 
 
 ## [0.9.0] 2021-09-03